requirements.txt
fastapi
uvicorn
pydantic
bcrypt
mysql-connector-python
sqlalchemy
alembic
pandas
numpy
scikit-learn
xgboost
joblib
python-multipart
python-dotenv
requests
passlib[bcrypt]
pyjwt
fastapi-mail
matplotlib
emails
tensorflow